Remove new method from PopulationTraits trait
The new method has been removed from the PopulationTraits trait definition as it's not necessary for this trait. This results in simplifying the trait's definition and not requiring implementing, what would likely be similar, new methods in each struct that implements this trait. The related documentation and implementation details have been cleaned up as well.

Also additional fixes for refactoring Population from vector to struct.